pyproject.toml
[tool.poetry]
name = "pythonstudy"
version = "0.1.0"
description = ""
authors = ["YunGyeongee <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
requests2 = "^2.16.0"
beautifulsoup4 = "^4.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
